mse/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
ultralytics==8.3.0
pyyaml==6.0.2
python-multipart==0.0.9
aiofiles==24.1.0
pytest==8.3.2